Refactor 'vector' module + Vector2D/3D traits
Splits the 'vector' module apart into submodules for 2D (xy) and
3D (xyz) vector types.

Additionally adds the `Vector2D` and `Vector3D` traits for these vector
types.
