+++ README.md
@@ -9,7 +9,13 @@
 [![Gitter Chat][gitter-image]][gitter-link]
 
 Generic Rust accelerometer support, including traits and types for taking
-readings from 2 or 3-axis accelerometers
+readings from 2 or 3-axis accelerometers.
+
+Includes minimal `no_std`-oriented 2D and 3D vector types useful for
+acceleration readings as well as mathematical operations on them.
+
+Additionally includes minimal statistical analysis tools for filtering
+noisy accelerometer data.
 
 [Documentation][docs-link]
